chore: refactor from controller to application
